Add new docs parser
